feat: [#46 - ] Comment 영역의 사용자 정보 수집 방식, 도메인 정책, DTO 필드 수정 - 요청 DTO의 불필요한 필드 제거 - controller 영역에서 인증 정보로부터 사용자 정보를 수집하도록 수정 - Comment에 대한 수정/삭제 정책 추가 Fixes: #46

+++ src/main/java/com/komentum/post/service/CommentService.java
@@ -2,13 +2,16 @@ package com.komentum.post.service;
 
 import com.komentum.post.domain.Comment;
 import com.komentum.post.domain.Post;
+import com.komentum.post.domain.policy.CommentPolicy;
 import com.komentum.post.dto.CommentDto.CommentCreateDto;
 import com.komentum.post.dto.CommentDto.CommentUpdateDto;
 import com.komentum.post.repository.CommentRepository;
 import com.komentum.user.domain.User;
+import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,14 +20,18 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
   private final CommentRepository commentRepository;
+  private final CommentPolicy commentPolicy;
 
+  @Transactional(readOnly = true)
   public List<Comment> getComments(Long postId, Pageable pageable) {
     return commentRepository.findAllByPost_PostId(postId, pageable);
   }
 
+  @Transactional(readOnly = true)
   public Comment getCommentById(Long commentId) {
     return commentRepository.findById(commentId)
-        .orElseThrow(() -> new RuntimeException("Comment not found"));
+        .orElseThrow(() -> new EntityNotFoundException(
+            String.format("failed to find post with id : %d", commentId)));
   }
 
   @Transactional
@@ -36,6 +43,9 @@ public class CommentService {
   @Transactional
   public Comment updateComment(Long commentId, CommentUpdateDto commentUpdateDto) {
     Comment comment = getCommentById(commentId);
+    if (!commentPolicy.canUpdate(comment.getUser())) {
+      throw new AccessDeniedException("failed to update comment : invalid user or role");
+    }
     comment.update(commentUpdateDto);
     return commentRepository.save(comment);
   }
@@ -43,6 +53,9 @@ public class CommentService {
   @Transactional
   public void deleteComment(Long commentId) {
     Comment targetComment = getCommentById(commentId);
+    if (!commentPolicy.canDelete(targetComment.getUser())) {
+      throw new AccessDeniedException("failed to delete comment : invalid user or role");
+    }
     commentRepository.deleteById(targetComment.getCommentId());
   }
 }
